[CR] Allow for decimal places in episode number
Addresses #615

+++ crunchy.ts
@@ -1382,7 +1382,7 @@ export default class Crunchy implements ServiceClass {
 
       variables.push(...([
         ['title', medias.episodeTitle, true],
-        ['episode', isNaN(parseInt(medias.episodeNumber)) ? medias.episodeNumber : parseInt(medias.episodeNumber), false],
+        ['episode', isNaN(parseFloat(medias.episodeNumber)) ? medias.episodeNumber : parseFloat(medias.episodeNumber), false],
         ['service', 'CR', false],
         ['seriesTitle', medias.seriesTitle, true],
         ['showTitle', medias.seasonTitle, true],
